requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.0.0
sendgrid>=6.11.0
supabase>=2.7.0
twilio>=9.3.0